athelo-api/requirements.txt
pg8000==1.29.8
cloud-sql-python-connector==1.2.4
gunicorn==20.1.0
Werkzeug==2.3.7
flask-smorest==0.42.1
alembic==1.12.0
flask-sqlalchemy==3.0.0
marshmallow-sqlalchemy==0.29.0